requirements.txt
flask==2.3.3
pandas==2.0.3
openpyxl==3.1.2
google-generativeai==0.3.1
werkzeug==2.3.7
nltk==3.9.1
python-dotenv==1.1.0
numpy==1.26.4